+++ lib/model.dart
@@ -12,7 +12,6 @@ import 'package:flutter/material.dart';
 import 'dart:async';
 import 'package:rxdart/rxdart.dart';
 import 'dart:typed_data';
-import 'package:json_serializable/json_serializable.dart';
 import 'package:json_annotation/json_annotation.dart';
 
 // Importing the UI library from Research Package