+++ lib/research_package.dart
@@ -1,0 +1,4 @@
+library research_package;
+
+export 'model.dart';
+export 'ui.dart';