+++ lib/src/model/result/RPTaskResult.dart
@@ -3,7 +3,13 @@ part of research_package_model;
 /// The result object a Task creates
 @JsonSerializable(fieldRename: FieldRename.snake, includeIfNull: false)
 class RPTaskResult extends RPResult {
-  Map<String, RPStepResult> _results;
+  /// The results in of a Task in a map.
+  ///
+  /// Each entry represents a Step's result ([RPStepResult]).
+  /// The keys are strings with the identifiers of the corresponding Step.
+  Map<String, RPStepResult> results;
+
+  RPTaskResult();
 
   /// Returns an [RPTaskResult] with a given identifier and an empty map of results.
   ///
@@ -11,23 +17,17 @@ class RPTaskResult extends RPResult {
   /// It sets [startDate] to the ```DateTime.now()```. Since these objects are instantiated
   /// together with the Task it belongs to so it can be used for measuring
   /// how much time the participant spent the given Task.
-  RPTaskResult(String identifier) : super(identifier) {
-    this._results = Map<String, RPStepResult>();
+  RPTaskResult.withParams(String identifier) : super.withIdentifier(identifier) {
+    this.results = Map<String, RPStepResult>();
     startDate = DateTime.now();
   }
 
-  /// The results in of a Task in a map.
-  ///
-  /// Each entry represents a Step's result ([RPStepResult]).
-  /// The keys are strings with the identifiers of the corresponding Step.
-  Map<String, RPStepResult> get results => _results;
-
   /// Returns result value for the given identifier from the [results] map
-  RPStepResult getStepResultForIdentifier(String identifier) => _results[identifier];
+  RPStepResult getStepResultForIdentifier(String identifier) => results[identifier];
 
   /// Adds a result to the result map with the given identifier.
   void setStepResultForIdentifier(String identifier, RPStepResult stepResult) {
-    _results[identifier] = stepResult;
+    results[identifier] = stepResult;
   }
 
   factory RPTaskResult.fromJson(Map<String, dynamic> json) => _$RPTaskResultFromJson(json);