+++ CHANGELOG.md
@@ -1,7 +1,17 @@
 ## 0.0.2
-* added support for serialization to/from JSON
+ * added initial support for serialization to/from JSON
+ * JSON serialization is available for these classes:
+    * `RPAnswerFormat`
+    * `RPChoiceAnswerFormat`
+    * `RPIntegerAnswerFormat`
+    * `RPConsentDocument`
+    * `RPConsentSection`
+    * `RPSignatureResult`
+    * `RPStepResult`
+    * `RPTaskResult`
+    * `RPChoice`
 
 ## 0.0.1
  * initial release
  * entire framework done
- * support for SingleChoice question format
+ * support for SingleChoice question type